livepatch-patch-hook: add support for livepatch sympos structures
Support for patching objects that have duplicated function names will be
introduced upstream in v4.5. This structure change needs to be handled by
the livepatch patch hook.

In addition, we need to be able to look up the symbol position in order to
populate old_sympos and sympos fields in klp_object and klp_reloc
respectively.

Signed-off-by: Chris J Arges <chris.j.arges@canonical.com>